+++ fixtures/embedded-model/calc-v1/test/calc/core_test.clj
@@ -1,0 +1,6 @@
+(ns calc.core-test
+  (:require [calc.core :as calc]
+            [clojure.test :refer [deftest is]]))
+
+(deftest square-three
+  (is (= 9 (calc/square 3))))